move extension methods to Predef instead of Scala3RunTime
