+++ library/src/scala/runtime/Scala3RunTime.scala
@@ -1,7 +1,5 @@
 package scala.runtime
 
-import scala.annotation.since
-
 object Scala3RunTime:
 
   // Called by inline def assert's. Extracted to minimize the bytecode size at call site.
@@ -20,13 +18,5 @@ object Scala3RunTime:
     val isNull = x == null
     if (isNull) throw new NullPointerException("tried to cast away nullability, but value is null")
     else x.asInstanceOf[x.type & T]
-
-  extension (inline x: AnyRef | Null)
-    @since("3.2")
-    inline def eq(inline y: AnyRef | Null): Boolean =
-      x.asInstanceOf[AnyRef] eq y.asInstanceOf[AnyRef]
-    @since("3.2")
-    inline def ne(inline y: AnyRef | Null): Boolean =
-      !(x eq y)
 
 end Scala3RunTime
